refactor(slideshow-form): update imports and enhance path aliases
- Removed vite.config.ts from .gitignore for better build integration.
- Added @types/node as a dev dependency for improved TypeScript support.
- Updated TypeScript path aliases for better module resolution across components and schemas.
- Cleaned up unused imports and streamlined component structure in the slideshow form.
- Enhanced the organization of the project by ensuring consistent import paths.
