+++ client/.gitignore
@@ -7,8 +7,6 @@ yarn-error.log*
 pnpm-debug.log*
 lerna-debug.log*
 
-vite.config.ts
-
 node_modules
 dist
 dist-ssr
