+++ client/package.json
@@ -40,6 +40,7 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.27.0",
+    "@types/node": "^22.15.29",
     "@types/react": "^19.1.4",
     "@types/react-dom": "^19.1.5",
     "@types/react-router-dom": "^5.3.3",
