+++ client/src/features/slideshow-form/components/Experience/AddExperienceButton.tsx
@@ -1,6 +1,6 @@
 import { Box, HStack } from "@chakra-ui/react";
-import BaseButton from "@/components/ui/BaseButton";
 import { FiPlus } from "react-icons/fi";
+import BaseButton from "@/components/ui/BaseButton";
 
 interface AddExperienceButtonProps {
   onAdd: () => void;
