+++ client/src/features/slideshow-form/dashboard/main-layout/section-cards/EducationCard.tsx
@@ -1,8 +1,8 @@
 import { MdSchool, MdVerifiedUser, MdCalendarToday } from "react-icons/md";
 import { Icon, Box, Stack, Text, Badge, HStack, Flex } from "@chakra-ui/react";
 import { useNavigate } from "react-router-dom";
-import { educationData, getCompletionColor } from "@/features/slideshow-form/utils/mockData";
-import { getSectionStepPath } from "@/features/slideshow-form/routes/slideshowRoutes";
+import { educationData, getCompletionColor } from "@slideshow-form/utils/mockData";
+import { getSectionStepPath } from "@slideshow-form/routes/slideshowRoutes";
 
 const EducationCard = () => {
   const navigate = useNavigate();
