+++ client/src/features/slideshow-form/dashboard/main-layout/section-cards/ExperienceCard.tsx
@@ -1,8 +1,8 @@
 import { MdWork, MdBusiness, MdDateRange } from "react-icons/md";
 import { Icon, Box, Stack, Text, Badge, HStack, Flex } from "@chakra-ui/react";
 import { useNavigate } from "react-router-dom";
-import { experienceData, getCompletionColor } from "@/features/slideshow-form/utils/mockData";
-import { getSectionStepPath } from "@/features/slideshow-form/routes";
+import { experienceData, getCompletionColor } from "@slideshow-form/utils/mockData";
+import { getSectionStepPath } from "@slideshow-form/routes";
 
 const ExperienceCard = () => {
   const navigate = useNavigate();
