+++ client/src/features/slideshow-form/dashboard/main-layout/section-cards/PreferencesCard.tsx
@@ -1,8 +1,8 @@
 import { MdSettings, MdLocationOn, MdAttachMoney, MdWork } from "react-icons/md";
 import { Icon, Box, Stack, Text, Badge, HStack, Flex } from "@chakra-ui/react";
 import { useNavigate } from "react-router-dom";
-import { preferencesData, getCompletionColor } from "@/features/slideshow-form/utils/mockData";
-import { getSectionStepPath } from "@/features/slideshow-form/routes";
+import { preferencesData, getCompletionColor } from "@slideshow-form/utils/mockData";
+import { getSectionStepPath } from "@slideshow-form/routes";
 
 const PreferencesCard = () => {
   const navigate = useNavigate();
