+++ client/tsconfig.app.json
@@ -7,8 +7,11 @@
     "module": "ESNext",
     "skipLibCheck": true,
     "moduleResolution": "Bundler",
+    "baseUrl": "./src",
     "paths": {
-      "@/*": ["./src/*"]
+      "@/*": ["./*"], // General alias for everything inside `src`
+      "@slideshow-form/*": ["./features/slideshow-form/*"], // Existing alias
+      "@@/features/*": ["./features/*"] // Alternative alias for features
     },
     "types": ["react"],
 
@@ -26,5 +29,5 @@
     "noFallthroughCasesInSwitch": true,
     "noUncheckedSideEffectImports": true
   },
-  "include": ["src"]
+  "include": ["src", "features"]
 }
